#123 Display Category.name, serviceAssignment.name, or searchBy on search results page based on what searchBy used.
This commit refactored the results view, including splitting out the rendered block into 3, conditional on how many results are possible.  This again, caused a change in how searching is done when the user provides a specific value. If a typed-in value, we assume the user knows exactly what they are looking for (eg serial number or barcode) and so search for that value exactly. Previously we would fuzzy search for it with regex against the requested field, but it was too fuzzy and searching for '00' returned every tool.
now, when a search result is rendered, it either looks up the category/SA from res.locals or it just shows the searchTerm and total found result.
Also in this commit, as it was adjacent, we added a res.locals value for totalFound. The "Total Results: xx' item in the results view could only see however many were on the page loaded, not the total found. This is been fixed.