Add forgotPassword view; reorganize auth templates
Introduced a new view for the forgot password functionality, enabling users to reset their passwords securely. This comes with a dedicated form that requires users to submit their new password and confirm it for verification purposes.

Additionally, restructured the authentication-related views by moving 'login.hbs' and 'register.hbs' into an auth directory. This change streamlines the project structure, making it easier to manage views related to authentication in a single, logical location.

These adjustments enhance the application's usability and maintainability, ensuring a smoother experience for both users attempting to recover access to their accounts and developers navigating the codebase.