feat: Add timestamps to Tool model
This commit adds the `timestamps` option to the `toolSchema` in the `Tool.model.js` file. By enabling this option, the `createdAt` and `updatedAt` fields will automatically be added to the schema and populated with the creation and update timestamps respectively. This change improves the tracking and management of tool records in the application.