feat: Add barcode field to updateTool function
The `updateTool` function in the `tool.js` middleware file has been modified to include a new `barcode` field. This change allows for updating the barcode of a tool when making a request to update the tool information.

Note: This commit message follows the established convention of using a prefix (feat) to indicate a new feature or enhancement.