feat: Add js invoked class for changed inputs
The code changes include adding a new class called "changed" to inputs that have been modified. This class is applied when the input value is changed, and it adds a background color of red and text color of white. This change improves the user experience by visually indicating which inputs have been modified.