fix: Update dependencies and add PropelAuth integration
NOT CURRENTLY WORKING, the cookie is getting passed back from propel auth but isn't getting assigned to req.user and idk why.

This commit updates the dependencies in the package.json file to include the latest versions of "@propelauth/express" and "@propelauth/javascript". It also integrates PropelAuth into the toolkeeper.js file by creating an authClient and hooking up buttons to redirect to the signup and login pages. The logout functionality is currently commented out.