Feat: Batch Create Tools (#246)
* initial form building based on newTool modal

* feature: Add batchCreateTools function and related routes and views. framing in progress for first half of process (user input)

* fix: added config options for prettier so handlebars partials are supported under the angular parser

* Updated tooltip copy

* feat: Add button for batch creating tools in newTool modal and continued working on form. currently adding new rows is working. Auto-incrementing is the currernt target

* added toastr package via CDN, using that for errors with NaN values when increment checked.
autoincrement works for number values only.
next task is incrementing the right-most numeric string in the value as a fallback (and specifically to support ToolID values as they exist

* feat: wrapped up front end stuff, moving on to request handling

* struggling with promises like I always do when i take a month off of coding anything meaningful.

* pulling batch creation workflow into a separate file

* feat: valid successful route working without issue.
todo: build out error handling for duplicate unique fields

* feat: Refactor batch tool import to handle duplicate fields

The code changes refactor the batch tool import process to handle duplicate fields. The `createBatchTool` function now checks for duplicates in the `serialNumber`, `barcode`, and `toolID` fields using separate functions. If a duplicate is found, a `DuplicateError` is thrown with the appropriate message and details. The error is then added to the `errorList` array. The `batchCreateTools` function filters out any undefined tools and fulfilled promises before returning the `newTools` and `errorList` arrays.

* feat: Add required attribute to barcode field in Tool model

Require barcode via db model and require it on batchCreate form

* chore: version bump

* Linting, anti-pattern removal, and documentation added

* chore: modified front-end js files to modules and modified HTML import to match

* trim empty tool objects and duplication errors from returned toolList before rendering results table

* removed extra logging