### Commit Message
#### Subject:
Remove redundant password hashing during reset

#### Body:
The password hash generation is redundant, as the password is already hashed prior to the update by the setter for that property.  Leaving this in would cause the password to be hashed twice when reset, but then failing to match properly when attempting future login requests.

#### References:
No specific issue references.